latest work additions
started working on the latest work section. placed images and sized
them, floated and ready to go. found a way to imitate the hover color
change on the images using opacity, but realized they have an overlay
making them appear darker than they are so will be poking around to
change that.